pyproject.toml
[build-system]
requires = ["poetry-core>=1.2.0b2"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "megan_global_explanations"
version = "0.1.0"
description = "Insert Description"
license = "MIT license"
authors = ["Jonas Teufel <jonseb1998@gmail.com>"]
maintainers = ["Jonas Teufel <jonseb1998@gmail.com>"]
readme = "README.rst"
keywords = []
packages = [
    { include = "megan_global_explanations" },
    { include = "megan_global_explanations/experiments"},
]
include = [
    "megan_global_explanations/VERSION",
]

[tool.poetry.scripts]
megan_global_explanations = 'megan_global_explanations.cli:cli'

[tool.poetry.dependencies]
python = ">=3.8.0,<4.0"
pycomex = ">=0.9.2"
click = ">=7.1.2"
jinja2 = ">=3.0.3"
matplotlib = ">=3.5.3"
numpy = ">=1.23.2"
python-decouple = ">=3.6"
poetry-bumpversion = ">=0.3.0"
graph-attention-student = ">=0.12.0"
visual-graph-datasets = ">=0.12.0"
umap-learn = ">=0.5.3"

[tool.poetry.dev-dependencies]
pytest = ">=7.1.3"
poetry-bumpversion = ">=0.3.0"

[tool.poetry_bumpversion.file."megan_global_explanations/VERSION"]
search = "{current_version}"
replace = "{new_version}"

[tool.poetry_bumpversion.file."README.rst"]
search = 'version-{current_version}-orange.svg'
replace = 'version-{new_version}-orange.svg'

[tool.poetry_bumpversion.file."megan_global_explanations/__init__.py"]
search = '__version__ = "{current_version}"'
replace = '__version__ = "{new_version}"'